requirements.txt
SQLAlchemy
fastapi
uvicorn
pydantic
mysqlclient
python-dotenv
